+++ mne/io/tests/test_compensator.py
@@ -26,16 +26,30 @@ def test_compensation():
     comp2 = make_compensator(raw.info, 3, 1, exclude_comp_chs=True)
     assert_true(comp2.shape == (311, 340))
 
+    # round-trip
+    desired = np.eye(340)
+    for from_ in range(3):
+        for to in range(3):
+            if from_ == to:
+                continue
+            comp1 = make_compensator(raw.info, from_, to)
+            comp2 = make_compensator(raw.info, to, from_)
+            assert_allclose(np.dot(comp1, comp2), desired, atol=1e-6)
+            assert_allclose(np.dot(comp2, comp1), desired, atol=1e-6)
+
     # make sure that changing the comp doesn't modify the original data
     raw2 = Raw(ctf_comp_fname).apply_gradient_compensation(2)
     assert_equal(get_current_comp(raw2.info), 2)
     fname = op.join(tempdir, 'ctf-raw.fif')
     raw2.save(fname)
     raw2 = Raw(fname)
-    assert_equal(get_current_comp(raw2.info), 3)
+    assert_equal(raw2.compensation_grade, 2)
+    raw2.apply_gradient_compensation(3)
+    assert_equal(raw2.compensation_grade, 3)
     data, _ = raw[:, :]
     data2, _ = raw2[:, :]
-    assert_allclose(data, data2, rtol=1e-9, atol=1e-20)
+    # channels have norm ~1e-12
+    assert_allclose(data, data2, rtol=1e-9, atol=1e-18)
     for ch1, ch2 in zip(raw.info['chs'], raw2.info['chs']):
         assert_true(ch1['coil_type'] == ch2['coil_type'])